+++ src/platform/linux.rs
@@ -1,6 +1,6 @@
 use std::error::Error;
 use std::process::Command;
-use Configuration;
+use configuration::Configuration;
 
 pub fn set_wallpaper(path: &str) -> Result<(), ()> {
     let mut command = Command::new("feh");
